test(engine/connection/exec): refactor exec lifecycle tests into separate modules
- Split the large exec lifecycle unit tests file into focused test files for attached, detached, error, and start options scenarios.
- Extracted shared test helpers and fixtures into a new helpers module for reuse.
- Improved test code organization and reduced the complexity of the main tests.rs file.
- Added missing test annotations and restructured mocks and fixtures for clearer test setup.
- Adjusted documentation and doc comments accordingly.

This refactor improves test maintainability without changing production code behavior.
