+++ tests/ui/stable_embedder_uses_only_stable_modules.rs
@@ -5,7 +5,7 @@
 //! Run via `cargo test` — trybuild compiles this file in isolation and fails
 //! the test if it does not compile.
 
-use podbot::api::{ExecContext, ExecMode, ExecRequest};
+use podbot::api::{CommandOutcome, ExecContext, ExecMode, ExecRequest, exec};
 use podbot::config::AppConfig;
 use podbot::error::Result;
 
@@ -15,6 +15,20 @@ fn _assert_exec_surface_is_importable(handle: &tokio::runtime::Handle) -> Result
         .with_tty(false);
     let config = AppConfig::default();
     ExecContext::connect(&config, handle)
+}
+
+fn _assert_top_level_exec_is_importable(
+    config: &AppConfig,
+    request: &ExecRequest,
+) -> Result<CommandOutcome> {
+    exec(config, request)
+}
+
+fn _assert_context_exec_is_importable(
+    context: &ExecContext,
+    request: &ExecRequest,
+) -> Result<CommandOutcome> {
+    context.exec(request)
 }
 
 fn main() {}
